Update compose monkey and fix compose bug
Compose monkey no longer immediately composes the initial document with
the first operation. Instead, it will compose all the operations
together and then compose the initial document with the result. This
simulates a more realistic scenario where the user might be buffering
operations. If the first operation is immediately composed with the
initial document, then you end up always doing a compose(insert, *).

As a result of this change, a bug was discovered where a trailing delete
in the first operation could cause an error. This is because of the case
where a delete in the first op does not have to be spanned by the second
op.