Import SHA1 function from libtomcrypt
I wanted to avoid using the entire libtomcrypt library because a) it's a
lot of code for a single hash function, and b) it would immediately make
the libot archive > 1MB in size.

The following changes were made to the original code:

* unsigned long types were changed to uint32_t.
* ulong32 and ulong64 types (specific to libtomcrypt) were changed to
  uint32_t and uint64_t, respectively.
* unsigned char types were changed to just char since libot builds with
  char as unsigned anyway.
* Any ifdefs originally used for build customization were removed.
* Platform-specific optimizations (endianness, asm, etc.) were removed
  in favor of their portable (and slightly slower) variants.
* Unused constants and test functions were removed.

The imported code can still be cleaned up some more. A lot of the
abstractions in the original library are unnecessary for a single hash
function. Adding some tests is also a must.