Change memory strategy for client buffer
Having the client's buffer or anticipated op point to an op in the
document was causing problems whenever a new op was appended to the
document. This is because the document is an array which can be
reallocated after an append, invalidating any pointers into the array.

This commit updates the client to perform a shallow copy of an op when
buffering, and then tracks if it needs to free the op's components or
not.