Set client_id and parent automatically
A client will set the client_id field of an op that it is applying. A
doc will set the parent field of an op if it's being appended.
Eventually, these two fields should be removed from the op constructor
since it makes more sense to set them when the op is actually being
used.