Add alen parameter to hextoa
This fixes a bug where a hex string shorter than the allocated array
would be incorrecty decoded. For example, if a has a length of 10 but
the hex string is "abc123", then the decoded array would end up looking
like "abc1230000" instead of "0000abc123". By knowing the length of the
array, we can correctly decode it as "0000abc123".