Fix bugs with buffer hashes
Client buffers were being sent with the incorrect hashes and parents.
This was causing the server to incorrectly try to transform ops instead
of composing them. This change fixes that by making sure buffers are
sent with the correct parent, and omitting the operation's hash.
Instead, the server is responsible for determining the hash of the
received buffer.

The reason why the client doesn't hash the buffer is because it would
need to know the server's latest state. While it's possible to have the
client track the server's state, it's faster and easier to just have the
server hash the incoming buffer.