Fix segfault with anticipated ops
A segfault could occur in cases where a client's buffer and anticipated
op shared the same components (because of a shallow copy). The
anticipated op's components would be accidentally freed when the buffer
was freed, causing xform_anticipated to behave unpredictably.

Also added a scenario test that tests for this situation.