Duplicate anticipated and buffered ops
Since a client's anticipated, buffered and latest ops could overlap, the
client would avoid duplicating them when it could. However, this made
memory management a lot more complicated. Before updating the buffered
or anticipated ops, we needed to do complex checks to see if the op was
being used somewhere else. Duplicating the ops is less efficient, but
simplifies the code tremendously.