Change ot_size so delete subtracts from the size
The size of an operation is used when determining if appending a new op
would cause a document to exceed its maximum size. When only inserts are
counted, then operations that replace text would incorrectly cause the
maximum limit to be reached. Having delete components subtract from the
size will give the correct result.